Create CountVowelsPermutation
This is taken from LeetCode daily challenge on the 4th of July. The challenge is to find the number of possible strings of vowels where: a can only be followed by e, e can be followed by a or i, i can be followed by everything except itself, 0 can only be followed by i or u, and u can only be followed by a. This is done using dynamic programing, bottom-up methodology.