Create ReduceArraySizeToHalf
Given an array of integers, find the minimum number of integers necessary to remove if we wish to reduce the array size to half of what it currently is. This means duplicate values will be removed. This will be accomplished by having an array which holds the number of times each value appears in the original array.