Create PushDominoes.py
LeetCode daily challenge for 7/21/2021. The task is to find out what will happen if we push a line of dominoes a certain way. If an upright domino is pushed from both directions, it does nothing. If it is pushed from one direction, it falls in the same direction it is being pushed. Note that some may be standing upright when it is all done.