Create PartitionArrayIntoDisjointIntervals.py
LeetCode daily challenge for 7/22/2021. The goal is to create two arrays. The first contains all values which are less than or equal to all values in the second array. The goal is to return the value where the second array begins. This uses two pointers to find the maximum value contained in the first array and the smallest value contained in the second array. As long as the second value is greater than the first, we know that the two arrays will fit the solution. This uses two pointers and iterates through the array once, so the time complexity is O(n) and the space complexity is O(k) since the memory used is limited only by the size of the array passed by the function. Everything else is fixed.